internal-docs-ai-assistant/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
asyncpg
redis[async]
sentence-transformers
qdrant-client
python-dotenv
psycopg2-binary
nltk
beautifulsoup4
python-magic
pandas
langchain
PyPDF2
docx
qdrant-client